refactor: community, study filter refactoring (#97)
