+++ src/constants/community.ts
@@ -1,4 +1,4 @@
-import type { FilterSelectedType } from "@/types/common";
+import type { FilterSelectedType } from "@/types/filter";
 import type { SelectOptionType } from "@/types/filter";
 
 export const POSITION_DATA = [
@@ -25,10 +25,6 @@ export const SORT_OPTION_DATA: SelectOptionType[] = [
   {
     value: "viewCount",
     label: "조회수 순",
-  },
-  {
-    value: "comment",
-    label: "댓글 순",
   },
 ];
